Fixed deletion of revision fields when approving
When we fetch a revision to compare the changes, it contains the fields and the fields it revises. Since the comparision code will see what's come from the user in terms of additions and removals, the service was deleting the canonical fields.

So when we are looping through the existing fields from the database to compare with the incoming changes we do an additional check to make sure the revisions match.